pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-sample"
authors = [
    {name = "Pete DeJoy", email = "pete@astronomer.io"},
    {name = "Josh Fell", email = "josh.d.fell@astronomer.io"},
]
license = {text = "Apache License 2.0"}
description = "A sample Apache Airflow provider package built by Astronomer."
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = ["apache-airflow>=2.4"]

[project.urls]
Homepage = "https://astronomer.io"
Source = "https://github.com/astronomer/airflow-provider-sample/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "sample_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "sample_provider.__version__"}

[tool.setuptools.packages.find]
exclude = ["*tests.*", "*tests"]
